Fix React #130 crash on user management table
Extract UserRow into a "use client" module so lucide-react icon
references are included in the client bundle. Turbopack was inlining
icons (e.g. ChartNoAxesColumnIcon, MegaphoneIcon) into the server
bundle only when they appeared in server-component JSX passed as
children to client components, leaving the client references as
undefined and crashing hydration with React error #130 → #310.
